[encoding] convert errors to SNAFU + more
- remove quick-error
- remove error module
- at text: EncodeTextError and DecodeTextError
- at encode/decode::basic, error is std::io::Error
- encode-level error + decode-level error types
- [parser] propagate error type differences
- [parser] move DicomElementMarker to an independent module
- [object] propagate error type differences
- [ul] propagate error type differences
- remove recursion_limit attributes

Misc:
- remove missing_docs warn lint
   - very annoying for error types
- refactor transfer syntax impls
   - split encoding and decoding
   - deprecate transfer_syntax specific modules

Fixes:

- add missing check on a write_all
