+++ encoding/Cargo.toml
@@ -20,7 +20,7 @@ inventory-registry = ['inventory']
 [dependencies]
 dicom-core = { path = "../core", version = "0.2.0" }
 dicom-dictionary-std = { path = "../dictionary-std", version = "0.2.0" }
-quick-error = "1.2.3"
 encoding = "0.2.33"
 byteordered = "0.5.0"
 inventory = { version = "0.1.4", optional = true }
+snafu = "0.6.8"
