+++ encoding/src/text.rs
@@ -18,12 +18,44 @@
 //!
 //! [`SpecificCharacterSet`]: ./enum.SpecificCharacterSet.html
 
-use crate::error::TextEncodingError;
 use encoding::all::{GB18030, ISO_8859_1, ISO_8859_2, ISO_8859_3, ISO_8859_4, ISO_8859_5, UTF_8};
 use encoding::{DecoderTrap, EncoderTrap, Encoding, RawDecoder, StringWriter};
+use std::borrow::Cow;
 use std::fmt::Debug;
+use snafu::{Backtrace, Snafu};
 
-type Result<T> = std::result::Result<T, TextEncodingError>;
+/// An error type for text encoding issues.
+#[derive(Debug, Snafu)]
+pub enum EncodeTextError {
+    /// A custom error message,
+    /// for when the underlying error type does not encode error semantics
+    /// into type variants.
+    #[snafu(display("{}", message))]
+    EncodeCustom {
+        /// The error message in plain text.
+        message: Cow<'static, str>,
+        /// The generated backtrace, if available.
+        backtrace: Backtrace,
+    }
+}
+
+/// An error type for text decoding issues.
+#[derive(Debug, Snafu)]
+pub enum DecodeTextError {
+    /// A custom error message,
+    /// for when the underlying error type does not encode error semantics
+    /// into type variants.
+    #[snafu(display("{}", message))]
+    DecodeCustom {
+        /// The error message in plain text.
+        message: Cow<'static, str>,
+        /// The generated backtrace, if available.
+        backtrace: Backtrace,
+    }
+}
+
+type EncodeResult<T> = Result<T, EncodeTextError>;
+type DecodeResult<T> = Result<T, DecodeTextError>;
 
 /// A holder of encoding and decoding mechanisms for text in DICOM content,
 /// which according to the standard, depends on the specific character set.
@@ -40,12 +72,12 @@ pub trait TextCodec {
     /// Decode the given byte buffer as a single string. The resulting string
     /// _may_ contain backslash characters ('\') to delimit individual values,
     /// and should be split later on if required.
-    fn decode(&self, text: &[u8]) -> Result<String>;
+    fn decode(&self, text: &[u8]) -> DecodeResult<String>;
 
     /// Encode a text value into a byte vector. The input string can
     /// feature multiple text values by using the backslash character ('\')
     /// as the value delimiter.
-    fn encode(&self, text: &str) -> Result<Vec<u8>>;
+    fn encode(&self, text: &str) -> EncodeResult<Vec<u8>>;
 }
 
 impl<T: ?Sized> TextCodec for Box<T>
@@ -56,11 +88,11 @@ where
         self.as_ref().name()
     }
 
-    fn decode(&self, text: &[u8]) -> Result<String> {
+    fn decode(&self, text: &[u8]) -> DecodeResult<String> {
         self.as_ref().decode(text)
     }
 
-    fn encode(&self, text: &str) -> Result<Vec<u8>> {
+    fn encode(&self, text: &str) -> EncodeResult<Vec<u8>> {
         self.as_ref().encode(text)
     }
 }
@@ -73,11 +105,11 @@ where
         (**self).name()
     }
 
-    fn decode(&self, text: &[u8]) -> Result<String> {
+    fn decode(&self, text: &[u8]) -> DecodeResult<String> {
         (**self).decode(text)
     }
 
-    fn encode(&self, text: &str) -> Result<Vec<u8>> {
+    fn encode(&self, text: &str) -> EncodeResult<Vec<u8>> {
         (**self).encode(text)
     }
 }
@@ -122,6 +154,18 @@ impl Default for SpecificCharacterSet {
 }
 
 impl SpecificCharacterSet {
+    /** Obtain the specific character set identified by the given code string.
+     *
+     * Supported code strings include the possible values
+     * in the respective DICOM element (0008, 0005).
+     *
+     * # Example
+     * 
+     * ```
+     * let character_set = SpecificCharacterSet::from_code("ISO_IR 100");
+     * assert_eq!(character_set, Some(SpecificCharacterSet::IsoIr100));
+     * ```
+     */
     pub fn from_code(uid: &str) -> Option<Self> {
         use self::SpecificCharacterSet::*;
         match uid.trim_end() {
@@ -182,16 +226,16 @@ macro_rules! decl_character_set {
                 $term
             }
 
-            fn decode(&self, text: &[u8]) -> Result<String> {
+            fn decode(&self, text: &[u8]) -> DecodeResult<String> {
                 $val
                     .decode(text, DecoderTrap::Call(decode_text_trap))
-                    .map_err(|e| TextEncodingError::new(e).into())
+                    .map_err(|message| DecodeCustom { message }.build())
             }
 
-            fn encode(&self, text: &str) -> Result<Vec<u8>> {
+            fn encode(&self, text: &str) -> EncodeResult<Vec<u8>> {
                 $val
                     .encode(text, EncoderTrap::Strict)
-                    .map_err(|e| TextEncodingError::new(e).into())
+                    .map_err(|message| EncodeCustom { message }.build())
             }
         }
     };
@@ -206,18 +250,18 @@ impl TextCodec for DefaultCharacterSetCodec {
         "ISO_IR 6"
     }
 
-    fn decode(&self, text: &[u8]) -> Result<String> {
+    fn decode(&self, text: &[u8]) -> DecodeResult<String> {
         // Using 8859-1 because it is a superset. Reiterations of this impl
         // should check for invalid character codes (#40).
         ISO_8859_1
             .decode(text, DecoderTrap::Call(decode_text_trap))
-            .map_err(|e| TextEncodingError::new(e).into())
+            .map_err(|message| DecodeCustom { message }.build())
     }
 
-    fn encode(&self, text: &str) -> Result<Vec<u8>> {
+    fn encode(&self, text: &str) -> EncodeResult<Vec<u8>> {
         ISO_8859_1
             .encode(text, EncoderTrap::Strict)
-            .map_err(|e| TextEncodingError::new(e).into())
+            .map_err(|message| EncodeCustom { message }.build())
     }
 }
